refactor sql queries and add delete query command
