+++ README.md
@@ -1,7 +1,7 @@
 # gearbot
 
 gearbot is a CLI application that checks a local music store, Hank's Music Exchange to see if a piece of music gear you want is available.
-The idea started from my desire to be on social media less. Now you won't have to waste time logging on to instagram daily to check for updates.
+Born out of my ever-growing frustration with using social media. Now you won't have to waste time logging on to instagram daily to check for updates.
 You'll just get notified as soon as the gear you want is avaiable!
 
 ## Requirements
