+++ src/cli.py
@@ -5,7 +5,7 @@ CLI app.
 import typer
 
 from app.commands import *
-from data.data import init_db
+from data.sql_funcs import init_db
 
 app = typer.Typer()
 
@@ -14,6 +14,7 @@ app.command("search")(add_gear_query)
 app.command("close-query")(close_gear_query)
 app.command("update-matches")(update_gear_matches)
 app.command("update-gear")(update_gear_list)
+app.command("delete")(delete_gear_query)
 
 if __name__ == "__main__":
     app()
