+++ src/data/update.py
@@ -1,15 +1,12 @@
 """
-Udate table data.
+Update table data.
 """
 
-from .data import execute_sql_query
+from .sql_funcs import sql_execute
 
 
 def update_query_status(search_term: str, is_open: bool):
     """Update the status of a query by SEARCHTERM."""
     params = (is_open, search_term)
-    sql_query = """UPDATE gear_queries
-        SET is_open = ?
-        WHERE search_term = ?
-    """
-    execute_sql_query(sql_query, params)
+    sql_query = """UPDATE gear_queries SET is_open = ? WHERE search_term = ?"""
+    sql_execute(sql_query, params)
